Add skylight style base and change the RSSLink
VPS has switched to using skylight as part of Pandoc.  So let's detect
that in the base.html.j2 layer.  Also, VPS is using Site.RSSLink instead
of Data.RSSLink (in a break with hugo, sadly).  So let's use that
instead when linking to our atom feed for the site.